@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
+    <TargetFramework>net10.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
@@ -9,7 +9,8 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.0.8" />
     <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.8" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.AzureAppConfiguration" Version="8.3.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.AzureAppConfiguration"
+      Version="8.3.0" />
     <PackageReference Include="MongoDB.Driver" Version="3.4.3" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="9.0.4" />
     <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="9.0.4" />
@@ -18,8 +19,10 @@
   <ItemGroup>
     <ProjectReference Include="..\Adapters\IBBS.AI.Agents.Adapters\IBBS.AI.Agents.Adapters.csproj" />
     <ProjectReference Include="..\Adapters\IBBS.API.Adapters\IBBS.API.Adapters.csproj" />
-    <ProjectReference Include="..\Adapters\IBBS.Infrastructure.MongoDB.Adapters\IBBS.Infrastructure.MongoDB.Adapters.csproj" />
-    <ProjectReference Include="..\Adapters\IBBS.Infrastructure.Persistence.Adapters\IBBS.Infrastructure.Persistence.Adapters.csproj" />
+    <ProjectReference
+      Include="..\Adapters\IBBS.Infrastructure.MongoDB.Adapters\IBBS.Infrastructure.MongoDB.Adapters.csproj" />
+    <ProjectReference
+      Include="..\Adapters\IBBS.Infrastructure.Persistence.Adapters\IBBS.Infrastructure.Persistence.Adapters.csproj" />
     <ProjectReference Include="..\Core\IBBS.Domain\IBBS.Domain.csproj" />
   </ItemGroup>
 
