@@ -1,4 +1,5 @@
 ﻿using IBBS.AI.Agents.Adapters.AgentManagers;
+using IBBS.AI.Agents.Adapters.Contracts;
 using IBBS.AI.Agents.Adapters.Helpers;
 using IBBS.Domain.DrivenPorts;
 using Microsoft.Extensions.Configuration;
@@ -12,35 +13,31 @@ namespace IBBS.AI.Agents.Adapters.IOC;
 /// </summary>
 public static class DIContainer
 {
-	/// <summary>
-	/// Adds the ai agents services.
-	/// </summary>
-	/// <param name="services">The services.</param>
-	/// <param name="configuration">The configuration.</param>
-	/// <returns>The service collection.</returns>
-	public static IServiceCollection AddAiAgentsServices(this IServiceCollection services, IConfiguration configuration) =>
-		services.AddScoped<IHttpClientHelper, HttpClientHelper>().AddScoped<IAiAgentsService, AiAgentsService>()
-			.ConfigureHttpClientFactory(configuration);
+    /// <summary>
+    /// Adds the ai agents services.
+    /// </summary>
+    /// <param name="services">The services.</param>
+    /// <param name="configuration">The configuration.</param>
+    /// <returns>The service collection.</returns>
+    public static IServiceCollection AddAiAgentsServices(this IServiceCollection services, IConfiguration configuration) =>
+        services.AddScoped<IHttpClientHelper, HttpClientHelper>().AddScoped<IAiAgentsService, AiAgentsService>()
+            .ConfigureHttpClientFactory(configuration);
 
-	// <summary>
-	/// Configures the HTTP client factory.
-	/// </summary>
-	/// <param name="services">The services.</param>
-	/// <param name="configuration">The configuration.</param>
-	private static IServiceCollection ConfigureHttpClientFactory(this IServiceCollection services, IConfiguration configuration)
-	{
-		services.AddHttpClient(ConfigurationConstants.AiAgentsHttpClient, client =>
-		{
-			var apiBaseAddress = configuration[ConfigurationConstants.AiAgentsApiBaseUrl];
-			if (string.IsNullOrEmpty(apiBaseAddress))
-			{
-				throw new ArgumentNullException(apiBaseAddress);
-			}
+    // <summary>
+    /// Configures the HTTP client factory.
+    /// </summary>
+    /// <param name="services">The services.</param>
+    /// <param name="configuration">The configuration.</param>
+    private static IServiceCollection ConfigureHttpClientFactory(this IServiceCollection services, IConfiguration configuration)
+    {
+        services.AddHttpClient(ConfigurationConstants.AiAgentsHttpClient, client =>
+        {
+            var apiBaseAddress = configuration[ConfigurationConstants.AiAgentsApiBaseUrl];
+            ArgumentException.ThrowIfNullOrEmpty(apiBaseAddress);
+            client.BaseAddress = new Uri(apiBaseAddress);
+            client.Timeout = TimeSpan.FromMinutes(3);
+        });
 
-			client.BaseAddress = new Uri(apiBaseAddress);
-			client.Timeout = TimeSpan.FromMinutes(3);
-		});
-
-		return services;
-	}
+        return services;
+    }
 }
