@@ -18,98 +18,98 @@ namespace IBBS.API.Adapters.Handlers;
 /// <seealso cref="IBBS.API.Adapters.Contracts.IAiServicesHandler" />
 public class AiServicesHandler(IAIService aiServices, IMapper mapper, IConfiguration configuration) : IAiServicesHandler
 {
-	/// <summary>
-	/// Generate the bug severity using LLM.
-	/// </summary>
-	/// <param name="bugSeverityAiRequest">The bug severity AI request model.</param>
-	/// <returns>The bug severity.</returns>
-	public async Task<string> GenerateBugSeverityAsync(BugSeverityAIRequestDTO bugSeverityAiRequest)
-	{
-		var domainRequest = mapper.Map<BugSeverityAIRequestDomain>(bugSeverityAiRequest);
-		return await aiServices.GenerateBugSeverityAsync(domainRequest).ConfigureAwait(false);
-	}
+    /// <summary>
+    /// Generate the bug severity using LLM.
+    /// </summary>
+    /// <param name="bugSeverityAiRequest">The bug severity AI request model.</param>
+    /// <returns>The bug severity.</returns>
+    public async Task<string> GenerateBugSeverityAsync(BugSeverityAIRequestDTO bugSeverityAiRequest)
+    {
+        var domainRequest = mapper.Map<BugSeverityAIRequestDomain>(bugSeverityAiRequest);
+        return await aiServices.GenerateBugSeverityAsync(domainRequest).ConfigureAwait(false);
+    }
 
-	/// <summary>
-	/// Generates the tag for story asynchronous.
-	/// </summary>
-	/// <param name="userName">The current user name.</param>
-	/// <param name="requestDTO">The story.</param>
-	/// <returns>
-	/// The genre tag response.
-	/// </returns>
-	public async Task<string> GenerateTagForStoryAsync(string userName, UserStoryRequestDTO requestDTO)
-	{
-		var domainRequest = mapper.Map<UserStoryRequestDomain>(requestDTO);
-		return await aiServices.GenerateTagForStoryAsync(userName, domainRequest).ConfigureAwait(false);
-	}
+    /// <summary>
+    /// Generates the tag for story asynchronous.
+    /// </summary>
+    /// <param name="userName">The current user name.</param>
+    /// <param name="requestDTO">The story.</param>
+    /// <returns>
+    /// The genre tag response.
+    /// </returns>
+    public async Task<string> GenerateTagForStoryAsync(string userName, UserStoryRequestDTO requestDTO)
+    {
+        var domainRequest = mapper.Map<UserStoryRequestDomain>(requestDTO);
+        return await aiServices.GenerateTagForStoryAsync(userName, domainRequest).ConfigureAwait(false);
+    }
 
-	/// <summary>
-	/// Gets the chatbot response asynchronous.
-	/// </summary>
-	/// <param name="chatMessageRequest">The user query request.</param>
-	/// <returns>
-	/// The ai agent response.
-	/// </returns>
-	public async Task<AIChatbotResponseDTO> GetChatbotResponseAsync(UserQueryRequestDTO chatMessageRequest)
-	{
-		var areFollowupQuestionsEnabled = bool.TryParse(configuration[ConfigurationConstants.AreFollowupQuestionsEnabled], out var parsedValue) && parsedValue;
+    /// <summary>
+    /// Gets the chatbot response asynchronous.
+    /// </summary>
+    /// <param name="chatMessageRequest">The user query request.</param>
+    /// <returns>
+    /// The ai agent response.
+    /// </returns>
+    public async Task<AIChatbotResponseDTO> GetChatbotResponseAsync(UserQueryRequestDTO chatMessageRequest)
+    {
+        var areFollowupQuestionsEnabled = bool.TryParse(configuration[ConfigurationConstants.AreFollowupQuestionsEnabled], out var parsedValue) && parsedValue;
 
-		var domainInput = mapper.Map<UserQueryRequestDomain>(chatMessageRequest);
-		var domainResponse = await aiServices.GetChatbotResponseAsync(domainInput, areFollowupQuestionsEnabled).ConfigureAwait(false);
-		return mapper.Map<AIChatbotResponseDTO>(domainResponse);
-	}
+        var domainInput = mapper.Map<UserQueryRequestDomain>(chatMessageRequest);
+        var domainResponse = await aiServices.GetChatbotResponseAsync(domainInput, areFollowupQuestionsEnabled).ConfigureAwait(false);
+        return mapper.Map<AIChatbotResponseDTO>(domainResponse);
+    }
 
-	/// <summary>
-	/// Gets the sample prompts for chatbot asynchronous.
-	/// </summary>
-	/// <returns>
-	/// The list of <see cref="LookupMasterDTO" />
-	/// </returns>
-	public async Task<IEnumerable<LookupMasterDTO>> GetSamplePromptsForChatbotAsync()
-	{
-		var domainResult = await aiServices.GetSamplePromptsForChatbotAsync().ConfigureAwait(false);
-		return mapper.Map<IEnumerable<LookupMasterDTO>>(domainResult);
-	}
+    /// <summary>
+    /// Gets the sample prompts for chatbot asynchronous.
+    /// </summary>
+    /// <returns>
+    /// The list of <see cref="LookupMasterDTO" />
+    /// </returns>
+    public async Task<IEnumerable<LookupMasterDTO>> GetSamplePromptsForChatbotAsync()
+    {
+        var domainResult = await aiServices.GetSamplePromptsForChatbotAsync().ConfigureAwait(false);
+        return mapper.Map<IEnumerable<LookupMasterDTO>>(domainResult);
+    }
 
-	/// <summary>
-	/// Moderates the content data asynchronous.
-	/// </summary>
-	/// <param name="userName">The current user name.</param>
-	/// <param name="requestDTO">The story.</param>
-	/// <returns>
-	/// The moderation content response.
-	/// </returns>
-	public async Task<string> ModerateContentDataAsync(string userName, UserStoryRequestDTO requestDTO)
-	{
-		var domainRequest = mapper.Map<UserStoryRequestDomain>(requestDTO);
-		return await aiServices.ModerateContentDataAsync(userName, domainRequest).ConfigureAwait(false);
-	}
+    /// <summary>
+    /// Moderates the content data asynchronous.
+    /// </summary>
+    /// <param name="userName">The current user name.</param>
+    /// <param name="requestDTO">The story.</param>
+    /// <returns>
+    /// The moderation content response.
+    /// </returns>
+    public async Task<string> ModerateContentDataAsync(string userName, UserStoryRequestDTO requestDTO)
+    {
+        var domainRequest = mapper.Map<UserStoryRequestDomain>(requestDTO);
+        return await aiServices.ModerateContentDataAsync(userName, domainRequest).ConfigureAwait(false);
+    }
 
-	/// <summary>
-	/// Posts the ai result feedback asynchronous.
-	/// </summary>
-	/// <param name="aiResponseFeedback">The ai response feedback.</param>
-	/// <param name="userEmail">The user email.</param>
-	/// <returns>
-	/// The boolean for success/failure.
-	/// </returns>
-	public async Task<bool> PostAiResultFeedbackAsync(AIResponseFeedbackDTO aiResponseFeedback, string userEmail)
-	{
-		var domainInput = mapper.Map<AIResponseFeedbackDomain>(aiResponseFeedback);
-		return await aiServices.PostAiResultFeedbackAsync(domainInput, userEmail).ConfigureAwait(false);
-	}
+    /// <summary>
+    /// Posts the ai result feedback asynchronous.
+    /// </summary>
+    /// <param name="aiResponseFeedback">The ai response feedback.</param>
+    /// <param name="userEmail">The user email.</param>
+    /// <returns>
+    /// The boolean for success/failure.
+    /// </returns>
+    public async Task<bool> PostAiResultFeedbackAsync(AIResponseFeedbackDTO aiResponseFeedback, string userEmail)
+    {
+        var domainInput = mapper.Map<AIResponseFeedbackDomain>(aiResponseFeedback);
+        return await aiServices.PostAiResultFeedbackAsync(domainInput, userEmail).ConfigureAwait(false);
+    }
 
-	/// <summary>
-	/// Rewrites with AI asynchronously.
-	/// </summary>
-	/// <param name="userName">The current user name.</param>
-	/// <param name="requestDTO">The story.</param>
-	/// <returns>
-	/// The AI response data
-	/// </returns>
-	public async Task<string> RewriteWithAIAsync(string userName, UserStoryRequestDTO requestDTO)
-	{
-		var domainRequest = mapper.Map<UserStoryRequestDomain>(requestDTO);
-		return await aiServices.RewriteWithAIAsync(userName, domainRequest).ConfigureAwait(false);
-	}
+    /// <summary>
+    /// Rewrites with AI asynchronously.
+    /// </summary>
+    /// <param name="userName">The current user name.</param>
+    /// <param name="requestDTO">The story.</param>
+    /// <returns>
+    /// The AI response data
+    /// </returns>
+    public async Task<string> RewriteWithAIAsync(string userName, UserStoryRequestDTO requestDTO)
+    {
+        var domainRequest = mapper.Map<UserStoryRequestDomain>(requestDTO);
+        return await aiServices.RewriteWithAIAsync(userName, domainRequest).ConfigureAwait(false);
+    }
 }
