@@ -9,11 +9,8 @@ var builder = WebApplication.CreateBuilder(args);
 builder.Configuration.SetBasePath(Directory.GetCurrentDirectory()).AddJsonFile(ConfigurationConstants.LocalAppsettingsFileConstant, optional: true).AddEnvironmentVariables();
 
 var credentials = builder.Environment.IsDevelopment()
-	? new DefaultAzureCredential()
-	: new DefaultAzureCredential(new DefaultAzureCredentialOptions
-	{
-		ManagedIdentityClientId = builder.Configuration[ConfigurationConstants.ManagedIdentityClientIdConstant]
-	});
+    ? new DefaultAzureCredential()
+    : new DefaultAzureCredential(new DefaultAzureCredentialOptions { ManagedIdentityClientId = builder.Configuration[ConfigurationConstants.ManagedIdentityClientIdConstant] });
 
 builder.Services.AddControllers();
 builder.Services.AddOpenApi();
@@ -21,19 +18,19 @@ builder.Services.AddCors(options => options.AddDefaultPolicy(policy => policy.Al
 
 builder.Services.AddSwaggerGen(options =>
 {
-	options.SwaggerDoc(SwaggerUIConstants.ApiVersion, new OpenApiInfo
-	{
-		Title = SwaggerUIConstants.ApplicationAPIName,
-		Version = SwaggerUIConstants.ApiVersion,
-		Description = SwaggerUIConstants.SwaggerDescription,
-		Contact = new OpenApiContact
-		{
-			Name = SwaggerUIConstants.AuthorDetails.Name,
-			Email = SwaggerUIConstants.AuthorDetails.Email
-		}
+    options.SwaggerDoc(SwaggerUIConstants.ApiVersion, new OpenApiInfo
+    {
+        Title = SwaggerUIConstants.ApplicationAPIName,
+        Version = SwaggerUIConstants.ApiVersion,
+        Description = SwaggerUIConstants.SwaggerDescription,
+        Contact = new OpenApiContact
+        {
+            Name = SwaggerUIConstants.AuthorDetails.Name,
+            Email = SwaggerUIConstants.AuthorDetails.Email
+        }
 
-	});
-	options.EnableAnnotations();
+    });
+    options.EnableAnnotations();
 });
 
 builder.ConfigureAzureAppConfiguration(credentials);
@@ -45,13 +42,13 @@ builder.Services.AddHttpContextAccessor();
 var app = builder.Build();
 if (app.Environment.IsDevelopment())
 {
-	app.MapOpenApi();
-	app.UseSwagger();
-	app.UseSwaggerUI(c =>
-	{
-		c.SwaggerEndpoint(SwaggerUIConstants.SwaggerEndpointUrl, $"{SwaggerUIConstants.ApplicationAPIName}.{SwaggerUIConstants.ApiVersion}");
-		c.RoutePrefix = SwaggerUIConstants.SwaggerUiPrefix;
-	});
+    app.MapOpenApi();
+    app.UseSwagger();
+    app.UseSwaggerUI(c =>
+    {
+        c.SwaggerEndpoint(SwaggerUIConstants.SwaggerEndpointUrl, $"{SwaggerUIConstants.ApplicationAPIName}.{SwaggerUIConstants.ApiVersion}");
+        c.RoutePrefix = SwaggerUIConstants.SwaggerUiPrefix;
+    });
 }
 
 app.UseExceptionMiddleware();
