@@ -1,8 +1,8 @@
-import { Environment } from "@/types/environment";
+import { Environment } from "@/app/Types/environment";
 
 export const environment: Environment = {
 	production: true,
-	apiBaseUrl: "https://your-production-api-url.com",
+	apiBaseUrl: "https://app-webapi-ibbs.azurewebsites.net/",
 	auth0Config: {
 		domain: "debanjans-lab.jp.auth0.com",
 		clientId: "QMzTORGKI2KXC4SZQ5EsRWUK6DVaJySO",
